refactor(api): move reconciliation auto route to prescription
