+++ api/tests/prescription/learner-management/integration/application/organization-learners-route_test.js
@@ -245,4 +245,54 @@ describe('Integration | Application | Organization Learners Management | Routes'
       expect(response.statusCode).to.equal(200);
     });
   });
+
+  describe('POST /api/sco-organization-learners/association/auto', function () {
+    const method = 'POST';
+    const url = '/api/sco-organization-learners/association/auto';
+
+    let httpTestServer;
+
+    beforeEach(async function () {
+      sinon
+        .stub(organizationLearnersController, 'reconcileScoOrganizationLearnerAutomatically')
+        .callsFake((request, h) => h.response('ok').code(204));
+      httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+    });
+
+    it('should succeed', async function () {
+      // given
+      const payload = {
+        data: {
+          attributes: {
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(204);
+    });
+
+    it('should return an error when there is an invalid campaign code attribute in the payload', async function () {
+      // given
+      const INVALID_CAMPAIGNCODE = '';
+      const payload = {
+        data: {
+          attributes: {
+            'campaign-code': INVALID_CAMPAIGNCODE,
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+  });
 });
